sr: move translations to config/localization
