+++ config/localization/translations/fr-fr.js
