+++ config/localization/translations/fr.js
