+++ config/localization/translations/index.js
